Correctly convert python dicts to be yaml compatible
With older versions of salt and python the string representation of
python dicts was close enough to json that they could be used directly
in yaml files. However this is now longer the case because objects like
strings might contain a u'' prefix which breaks compatibility. To avoid
these problems we explicitly convert python dicts to json when
outputting to the state file.